+++ admin/core/Admin.php
@@ -1345,7 +1345,10 @@ class Admin extends Origin
 	 */
 	public function printTableList(array $tables): void
 	{
-		$menuClass = ($this->settings->isNavigationDual() ? "class='dual'" : ($this->settings->isNavigationReversed() ? "class='reversed'" : ""));
+		$dualLinks = $this->settings->isNavigationDual() || $this->settings->isNavigationHover();
+
+		$menuClass = ($dualLinks ? "class='dual" . ($this->settings->isNavigationHover() ? " hover" : "") . "'" :
+			($this->settings->isNavigationReversed() ? "class='reversed'" : ""));
 
 		echo "<nav id='tables'><menu $menuClass>";
 
@@ -1371,7 +1374,7 @@ class Admin extends Origin
 
 				echo "<a href='$selectUrl'", bold($active, $class), " data-primary='true' title='$name'>$name</a>";
 
-				if ($this->settings->isNavigationDual() && $supportStructure) {
+				if ($dualLinks && $supportStructure) {
 					echo " <a href='$tableUrl' title='", lang('Show structure'), "' class='secondary'>", icon_solo("structure"), "</a>";
 				}
 			} else {
@@ -1385,7 +1388,7 @@ class Admin extends Origin
 					echo "<span data-primary='true'", bold($active, $class), ">$name</span>";
 				}
 
-				if ($this->settings->isNavigationDual()) {
+				if ($dualLinks) {
 					echo " <a href='$selectUrl' title='", lang('Select data'), "' class='secondary'>", icon_solo("data"), "</a>";
 				}
 			}
@@ -1407,6 +1410,7 @@ class Admin extends Origin
 				"" => lang('Default'),
 				Config::NavigationSimple => lang('Simple'),
 				Config::NavigationDual => lang('Dual'),
+				Config::NavigationHover => lang('Dual on hover'),
 				Config::NavigationReversed => lang('Reversed')
 			];
 			$default = $options[$this->config->getNavigationMode()];